Schema v8: int32-ms storage for position and spawn/death streams
MVD wire format encodes time as a 1-byte millisecond delta per
message. The decoder previously converted to float64 seconds on every
accumulation (mvd-reader/mvd/decoder.go:69, d.time += float64(td) /
1000.0), then narrowed again through []float32 when persisted as
PositionTrack.T. 1/1000 isn't representable in base-2 float, so the
running sum drifted; values that should have been exact integer ms
landed at e.g. 658.278992 instead of 658.279000.

That drift broke the boundary comparison boundaries[bIdx] <= t in
analyzer/locgraph.go and the blip filter in analyzer/timeline_streams.go,
where boundaries was built from Spawns/Deaths (float64 seconds) and t
from PositionTrack.T (float32 seconds). On gib-respawn cases (e.g.
e1m2 with a player dying at MH.low and instantly respawning at start)
the cursor failed to reset and locgraph emitted a spurious MH.low →
start teleport edge.

Fix: store and compare time as int32 milliseconds at the storage and
comparison boundary. Same 4 bytes per sample as the previous float32,
exact comparisons, no eps tolerance.

  - mvd-reader/mvd/decoder.go: canonical timeMs int32 accumulator,
    CurrentTimeMs(); CurrentTime() now derives float64(timeMs)*0.001.
    DemoMessage.TimeMs int32 alongside the existing Time float64.
  - mvd-reader/parser/{position,stats}.go: TimeMs int32 added to
    PlayerPositionEvent, SpawnEvent, DeathEvent; parser dispatcher
    plumbs msg.TimeMs through parsePlayerInfo / parseUpdateStat /
    parseUpdateStatLong / updateStat.
  - mvd-analytics/result/streams.go: PositionTrack.T []int32 (ms),
    PlayerStream.Spawns/Deaths []int32 (ms). JSON keys unchanged;
    consumers reading as seconds must scale by 1/1000.
  - mvd-analytics/result/result.go: CurrentSchemaVersion 7 -> 8 with
    a v8 history block.
  - mvd-analytics/analyzer/{state,timeline_streams,timeline,locgraph,
    tracks,postprocess}.go: streamBuilder posT/spawns/deaths []int32;
    recordPosition/Spawn/Death take int32 ms; mergeBoundaries +
    filterPositionLiBlips compare int32 directly; postprocess shifts
    via shiftAndFilterInts and a single int32(math.Round(start*1000))
    conversion.
  - mvd-analytics/view/{buckets,events,streamslice}.go: public window
    API stays float64 seconds, internal comparisons against the
    schema-v8 fields done in int32 ms after a single conversion at
    the entry point. ViewBucket.T stays float64 seconds so
    mvd-web/static/app.js is untouched.
  - mvd-api/handlers_test.go: X-Schema-Version expectation bumped.

Other timestamped fields (ChangeI16.T, Interval.Start/End,
MatchEvent.Time, frag/powerup event times) stayed float64 seconds —
they don't participate in the boundary comparison that motivated this
change. Pros/cons of finishing the migration in a future schema bump
captured in mvd-analytics/INT_MS_FOLLOWUPS.md.

Goldens regenerated; the diffs are exactly the float-precision drift
this change eliminates (e.g. 8.999999999956 -> 9). Docs updated:
RESULT_SCHEMA.md (PositionTrack + Spawns/Deaths rows, v8 description),
top-level README (schema reference + paragraph on the migration),
mvd-reader/MVD_FORMAT.md (notes the int32-ms accumulator in the
decoder), FOLLOWUPS.md (cross-link to INT_MS_FOLLOWUPS.md).

+++ mvd-analytics/result/streams.go
@@ -55,9 +55,12 @@ type PlayerStream struct {
 	Rockets []ChangeI16 `json:"rk,omitempty"`
 	Cells   []ChangeI16 `json:"cl,omitempty"`
 
-	// Discrete event timestamps (no value).
-	Spawns []float64 `json:"sp,omitempty"`
-	Deaths []float64 `json:"d,omitempty"`
+	// Discrete event timestamps (no value). Integer milliseconds since
+	// the stream's time origin (the same epoch as match-relative seconds
+	// elsewhere; schema v8 changed the type and unit to give exact
+	// comparisons against PositionTrack.T — see PositionTrack comment).
+	Spawns []int32 `json:"sp,omitempty"`
+	Deaths []int32 `json:"d,omitempty"`
 }
 
 // GlobalStream carries match-window anchors so consumers can resolve
@@ -72,6 +75,20 @@ type GlobalStream struct {
 // five arrays. Coordinates are int32 — Quake maps can exceed ±32 768
 // in any axis, so int16 would silently truncate.
 //
+// T is integer milliseconds since the stream's time origin (the same
+// epoch as the float-seconds version it replaced in schema v8). The
+// JSON key stayed "t" for compactness; consumers that previously read
+// it as seconds must scale by 1/1000 — the schema-version bump is the
+// signal. The wire format gives us a 1-byte ms delta per message, so
+// integer-ms storage keeps that exact value all the way from the
+// decoder through the persistence layer; float seconds reintroduced a
+// 1e-6 drift that caused spawn/death-boundary comparisons in locgraph
+// and the blip filter to land on the wrong side of an edge.
+//
+// Range: int32 ms gives ±24.8 days. Demos run minutes to hours, so
+// overflow isn't a concern. Negative values are valid after the
+// post-processor subtracts matchStart (warmup samples shift below 0).
+//
 // Li is the resolved loc-name index per native-rate sample (indexes
 // into TimelineAnalysisResult.LocTable, with 0 = "no loc"). Populated
 // during analyzer Finalize (after the loc finder is loaded), then
@@ -80,11 +97,11 @@ type GlobalStream struct {
 // view.Buckets — read this column directly instead of deriving locs
 // from x/y/z separately.
 type PositionTrack struct {
-	T  []float32 `json:"t"`
-	X  []int32   `json:"x"`
-	Y  []int32   `json:"y"`
-	Z  []int32   `json:"z"`
-	Li []int16   `json:"li,omitempty"`
+	T  []int32 `json:"t"` // milliseconds since the stream's time origin
+	X  []int32 `json:"x"`
+	Y  []int32 `json:"y"`
+	Z  []int32 `json:"z"`
+	Li []int16 `json:"li,omitempty"`
 }
 
 // ChangeI8 is a single transition in an int8 stream.
